Acceptance tests added  (#35)
* Acceptance tests added

* Stubbing moved to acceptance

* Wiremock removed from e2e env?

* Misspelling fix

* Fix for e2n

* Modal test removed

---------

+++ automated_testing/mock-end.sh
@@ -1,5 +1,11 @@
 echo "Stop MongoDB..."
 brew services stop mongodb-community@6.0
 
-echo "Stop WireMock..."
-pkill -f wiremock
+
+
+if [ "$E2E_ENV" = true ]; then
+    echo "Stop WireMock..."
+    pkill -f wiremock
+else
+    echo "E2E_ENV is not set to true. Skipping Wiremock execution."
+fi
